Showcases: default sleeve toggle to short on every listing
Cornerstone, Warpaint, and Unbreakable now initialize the sleeve selector to
short sleeve (the lower default price) instead of long. Unbreakable derives
DEFAULT_SLEEVE from HAS_SHORT so single-length collections still resolve correctly.

+++ src/components/CornerstoneShowcase.tsx
@@ -80,7 +80,7 @@ export default function CornerstoneShowcase() {
   const { addItem, openCart } = useCart();
   const prefersReducedMotion = useReducedMotion();
 
-  const [sleeve, setSleeve] = useState<SleeveLength>("long");
+  const [sleeve, setSleeve] = useState<SleeveLength>("short");
   // Single source of truth: derive activeVariant + activeIndex from these.
   // Default the configurator to Black when a Black variant exists.
   const defaultVariant =
